refactor(cli): lift bearer-supplier builder to shared/trpc
`instance/compose.ts` and `template/compose.ts` each had a verbatim copy
of `classifyTokenProviderError` plus a closure that calls the token
provider, maps `auth-required` to a typed Result, and throws otherwise.
Future auth-error classification changes would have had to land in both
places.

Extracts `createBearerSupplier(tokenProvider, host)` into
`shared/trpc/bearer-supplier.ts` and replaces both compose-time copies
with a single call. No behavior change.

Surfaced by automated review on #203.
