+++ packages/cli/src/modules/instance/compose.ts
@@ -3,7 +3,7 @@ import type { TokenProvider } from "../auth/index.js";
 import type { CompatService, ConfigService } from "../cli/index.js";
 import type { TemplateService } from "../template/index.js";
 import { createTrpcClient, type TrpcClient } from "../shared/trpc/trpc-client.js";
-import type { Result } from "../../result.js";
+import { createBearerSupplier } from "../shared/trpc/bearer-supplier.js";
 import { buildCreateCommand } from "./commands/create.js";
 import { buildDeleteCommand } from "./commands/delete.js";
 import { buildGetCommand } from "./commands/get.js";
@@ -13,7 +13,6 @@ import {
   createInstanceService,
   type InstanceService,
 } from "./services/instance-service.js";
-import type { AuthRequiredError } from "./domain/errors.js";
 
 /**
  * Composition options for the `instance` module.
@@ -54,20 +53,8 @@ export function composeInstanceModule(opts: InstanceModuleOptions): InstanceModu
   // typed `InstanceService` and the raw trpc client used by the
   // orchestration verbs (`create`, Phase 4 `delete` / `restart`) reuse
   // it so retries and refreshes stay consistent.
-  function getTokenFor(host: string): () => Promise<Result<string, AuthRequiredError>> {
-    return async () => {
-      const result = await opts.tokenProvider.getValidAccessToken(host);
-      if (result.ok) return result;
-      const classified = classifyTokenProviderError(result.error);
-      if (classified.kind === "auth-required") {
-        return { ok: false, error: classified };
-      }
-      throw new Error(classified.reason);
-    };
-  }
-
   const buildTrpc = (host: string): TrpcClient =>
-    createTrpcClient({ host, getToken: getTokenFor(host) });
+    createTrpcClient({ host, getToken: createBearerSupplier(opts.tokenProvider, host) });
 
   const createService = (host: string): InstanceService =>
     createInstanceService({ trpc: buildTrpc(host) });
@@ -125,37 +112,4 @@ export function composeInstanceModule(opts: InstanceModuleOptions): InstanceModu
     commands: [parent],
     exports: { createService },
   };
-}
-
-interface ReasonBearing {
-  reason?: string;
-  host?: string;
-  kind: string;
-}
-
-type ClassifiedError =
-  | { kind: "auth-required"; reason: string }
-  | { kind: "non-auth"; reason: string };
-
-/** Classify a `TokenProviderError` into two buckets without importing the
- *  auth domain's discriminant.
- *
- *  Only `not-logged-in` and `session-expired` route to `auth-required` —
- *  those are the cases where `dam auth login` is the fix. Everything else
- *  (refresh failures, auth-store I/O errors, malformed auth store) is a
- *  non-auth condition that login can't repair; the service layer surfaces
- *  those as `transport` errors carrying the original reason. */
-function classifyTokenProviderError(e: unknown): ClassifiedError {
-  if (typeof e !== "object" || e === null) {
-    return { kind: "non-auth", reason: "auth failure" };
-  }
-  const re = e as ReasonBearing;
-  switch (re.kind) {
-    case "not-logged-in":
-      return { kind: "auth-required", reason: re.host ? `not logged in to ${re.host}` : "not logged in" };
-    case "session-expired":
-      return { kind: "auth-required", reason: re.host ? `session expired for ${re.host}` : "session expired" };
-    default:
-      return { kind: "non-auth", reason: re.reason ?? re.kind };
-  }
 }
