+++ packages/cli/src/modules/template/compose.ts
@@ -2,6 +2,7 @@ import { Command } from "commander";
 import type { TokenProvider } from "../auth/index.js";
 import type { CompatService, ConfigService } from "../cli/index.js";
 import { createTrpcClient } from "../shared/trpc/trpc-client.js";
+import { createBearerSupplier } from "../shared/trpc/bearer-supplier.js";
 import { buildListCommand } from "./commands/list.js";
 import {
   createTemplateService,
@@ -26,15 +27,7 @@ export function composeTemplateModule(opts: TemplateModuleOptions): TemplateModu
   const createService = (host: string): TemplateService => {
     const trpc = createTrpcClient({
       host,
-      getToken: async () => {
-        const result = await opts.tokenProvider.getValidAccessToken(host);
-        if (result.ok) return result;
-        const classified = classifyTokenProviderError(result.error);
-        if (classified.kind === "auth-required") {
-          return { ok: false, error: classified };
-        }
-        throw new Error(classified.reason);
-      },
+      getToken: createBearerSupplier(opts.tokenProvider, host),
     });
     return createTemplateService({ trpc });
   };
@@ -56,29 +49,4 @@ export function composeTemplateModule(opts: TemplateModuleOptions): TemplateModu
     commands: [parent],
     exports: { createService },
   };
-}
-
-interface ReasonBearing {
-  reason?: string;
-  host?: string;
-  kind: string;
-}
-
-type ClassifiedError =
-  | { kind: "auth-required"; reason: string }
-  | { kind: "non-auth"; reason: string };
-
-function classifyTokenProviderError(e: unknown): ClassifiedError {
-  if (typeof e !== "object" || e === null) {
-    return { kind: "non-auth", reason: "auth failure" };
-  }
-  const re = e as ReasonBearing;
-  switch (re.kind) {
-    case "not-logged-in":
-      return { kind: "auth-required", reason: re.host ? `not logged in to ${re.host}` : "not logged in" };
-    case "session-expired":
-      return { kind: "auth-required", reason: re.host ? `session expired for ${re.host}` : "session expired" };
-    default:
-      return { kind: "non-auth", reason: re.reason ?? re.kind };
-  }
 }
